js: export automerge/next as a subpath export
Problem: when using the `next` API it is awkward to do things like this:

```javascript
import { next } from "@automerge/automerge"

const { init } = next
```

Even more problematic is importing types from the `next` package in
typescript.

Solution: add a ["subpath export"][1] for "@automerge/automerge/next".
This means that the above code looks like this:

```javascript
import { init } from "@automerge/automerge/next"
```

Actually getting this to work involved a fair amount of surgery to make
most of the code into ESM modules and make the nodejs package a commonjs
module by renaming it to `automerge_wasm.cjs`. All of this is pretty
mechanical though - a matter of changing import names and filenames. The
logic has remained the same.

[1]: https://nodejs.org/api/packages.html#subpath-exports