Fix missing marks at end of text (#720)
Problem: Applying a change which adds a mark to the last character in a
text object would not generate a patch for the change in question. The
reason this occurs is to do with the fact that  marks are implemented as
two consecutive operations, a `MarkBegin` op and a `MarkEnd` op. The
logic to determine whether to generate a patch is implemented as a
`TreeQuery` (specifically, as `SeekMark`). The `TreeQuery` gets called for
every element in the text sequence and scans through looking for the last
`MarkBegin` of the same type as the target `MarkEnd`, and then counts indices
until the insertion index is encountered. This fails if the insertion index is
at the end of the sequence because the `TreeQuery` never gets called.

Solution: Add a `SeekMark::finish` method which checks if the end mark
was ever found and if it wasn't assumes that the mark ended at the last
index of the sequence.