js: Transpile ESM to CommonJS (#731)
Problem: Previously, in 5f43505bc88d8789b9279c88505280559cb83443 I modified the
typescript config to output commonjs modules for commonjs consumers.
However, the module format being correct doesn't actually fix things
because when loading the module the loader (well, node at least) notices
that the package the cjs output is in has `"type": "module"` and so
attempts to load the JS as ESM modules and chokes.

Solution: Output the js files in `dist/cjs` with a `*.cjs` suffix. This
lets node know that these files are definitely absolutely commonJS
modules and please load them as commonjs modules, thank you.
Unfortunately `tsc` won't do this for us as it requires not just
renaming the output files, but also rewriting the import paths in the
transformed files to havea a `*.cjs` suffix and typescript has a design
principle of never rewriting imports. This means that we instead use
`esbuild` to transform the ESM output to CJS.