Add functionality to convert strings to text on load
Problem: As people transition to the new `next` JavaScript API they need
to handle old data which is using scalar strings instead of text
sequences. The way this manifests is that data which was a `string` in
the old API appears as a `RawString` when loaded using the new API. This
means that people have to either write code which handles both
`RawString` and `string`, or write code which converts the `RawString`
to `string`.

Solution: Offer an option to automatically convert the old strings
(which are `ScalarValue::Str` in Rust) to text sequences when loading
the document. This is achieved by actually adding a change to the
document setting all visible string values to text values with the same
contents. This isn't a complete solution because if you are syncing with
someone who is still using the old version of the application you will
end up flip flopping between the value the new application sees and the
value the old application sees. However, having the option is still
useful for applications who can coordinate some kind of data migration.