js: check arguments to diff are actually arrays
Problem: the typescript types for diff check that the arguments `before`
and `after` are arrays of strings, but if the user ignores those types
(or is using javascript) then we don't actually check at runtime that
they are valid. These arrays get passed through to the wasm, which can
explode messily.

Solution: throw a nice exception in `diff` so that at least the user can
find out where they went wrong