javascript: Better errors when setting invalid values (#795)
Context: The change callback which is used to modify an automerge
document is passed a proxy which looks like a POJO but intercepts
mutating calls and converts them into operations on the underlying
document.

Problem: When setting a key of one of the proxies in the change handler
to `undefined` the proxy throws a RangeError. This is necessary because
automerge doesn't support `undefined` - only `null`. However, the
`RangeError` doesn't contain any information which allows the developer
to figure out where they passed in an invalid value.

Solution: Include a JSON pointer to the path the developer attempted to
insert the invalid value at so they can identify what part of their data
they need to change.