Always send at least one sync message
Context: The sync protocol maintains a sync state which is updated every
time we send or receive a message. This sync state contains a field for
the heads the other end has reported. This is useful for applications to
understand what local changes the other hand has not yet acknowledged.

Problem: In some cases we never send a response to the first message in
the sync protocol and so the other end never learns what our heads are.
The reason this happens is that we have some optimisations in the
protocol to not send a message if we can tell that the other end has all
the same changes as us. One scenario where this happens is if you load a
document you previously saved and then synchronise that document with a
sync server which already has the document and there have been no new
changes.

Solution: Add a field to the sync state to track whether we have ever
sent a response and ensure we always send a response if we have not sent
one.