sync: send the whole document in the first sync message (#803)
Problem: when synchronizing a document for the first time the peer who
has the document sends a message containing all the changes in the
document encoded individually. For documents with a large number of
changes this is much slower to load than the compressed document format
of the same changes.

Solution: if we detect that the sending peer has no changes at all for
the document we are syncing then send the compressed document format
instead of the individual changes.

This is a fairly large patch for such a conceptually simple change. The
reason is that the wire format for the sync protocol was not designed
with forward compatibility in mind and so a little fiddling is
necessary. The first message that each peer sends has appended to it a
set of "capabilities", which are flags indicating what kinds of messages
the sending peer can accept. For old peers this will be empty, but for
new peers this will contain a flag indicating that a "v2" message -
which can contain the whole document - may be processed.