Introduce Automerge.updateText (rust: Transactable::update_text)
Problem: updating text fields in an automerge documents requires using
splice, which is intended for use when capturing individual input
events as the user creates them. In some environments it is impossible
to capture these events, leaving the developer to try and figure out a
set of splice calls. This is reasonably fiddly work.

Solution: introduce Transactable::update_text (which is exposed in
javascript as Automerge.updateText). This method performs a myers diff
to calculate a minimal set of edits to make based on the new text which
is passed to it.

The implementation is copied from the `similar` crate, which is licensed
under the Apache license. This necessitates including the Apache 2
license alongside the cribbed code as well as attribution comments in
the (2) files which were copied. This is preferable to pulling in the
entire `similar` crate because similar is quite large and we only need a
few hundred lines from it.