Add a clock argument to MarkMap::process (#841)
Problem: In 70438774f2f8a52a1346458a968249b3634696bc I fixed a bug where
marks were not being correctly calculated when a clock was passed to the
`nth` query. This was actually a special case of a more general problem
where we were not taking account of clocks passed to various queries
when processing marks.

Solution: Add an `Option<Clock>` argument to `MarkMap::process` and do the
visibility check in `process` rather than in the caller.