Fix a panic when rolling back a transaction
Context: The logic to rollback a transaction removed the actor ID of the
current actor if actor had no ops in the document. The purpose of this
was to stop actors with no ops in the OpSet from being written to the
document when calling `Automerge::save`. Also of note, actor IDs in the
document are not references to `ActorId`s, but indexes into a cached array
of `ActorId`s.

Problem: the logic to remove the actor ID contained a check that the
actor ID being removed was the last actor in the cached array of actor
IDs. This was not always true. For example, committing an empty
transaction, then merging changes from another peer would mean that the
final index in the cache was for a different actor ID to the current
actor.

Solution: Instead of attempting to remove the actor ID on rollback, just
filter the actor IDs when we save to ensure that no actors without ops
are saved. This is more robust but comes at the cost of an additional
scan over the `OpSet` when saving.