precompute visible paths when generating patches
Problem: in documents with histories that create a large number of
objects patch generation can become slow due to the cost of computing
the path for each path. We don't store the path anywhere in the
document, instead each OpTree stores the ID of op which created it; this
in turn means that to compute the path to an object we recursively seek
through the opset for each object in the path. In documents which have a
large number of ops - and especially if these ops are split over many
small objects - these seek operations can take a long time.

Solution: when we think we have more than a few paths to generate (this
patch arbitrarily choosese 100 paths) first precompute all the visible
paths in the document and then use this lookup table. To put this in
concrete terms, I have a document which prior to this patch took 50
seconds to diff and which after this patch takes 250ms to diff.

The implementation works by adding `Automerge::visible_object_paths`
which returns a map from object ID to path. This can be passed a set of
heads to generate the visible paths as at some point in time. In order
to make passing the to the `PatchLog` cleaner we add a `ReadDocInternal`
trait which exposes a `visible_object_paths` method, this trait is
implemented by both `Automerge` and `ReadDocAt`. We paramterise the
`PatchBuilder` on a `ReadDocInternal` implementation which means that we
can pass the `ReadDocInternal` into the `PatchBuilder` constructor and
then be sure that we are using the same paths throughout by using the
parameterised `ReadDocInternal` rather than passing a `ReadDoc` in to
each method on `PatchBuilder`.