rust/cli: emit text as string in JSON
Problem: the JSON formatting which automerge-cli uses emitted text
sequences as arrays of characters. This is counterintuitive as the JS
implementation now uses javascript strings and it's also irritating to
read.

Solution: switch to emitting text as a JSON string. Also clean up the
implementation by using the `automerge::AutoSerde` implementation of
`serde::Serialize` to serialize the doc to `serde_json` rather than the
custom implementation in `automerge-cli`.