include actor IDs from missing changes in save
Problem: in f10981 I modified the logic of `Automerge::save()` to do a
scan over the ops in the document when determining which actors to save.
This logic fails to account for empty commits by actors who never make
any ops.

Solution: rather than scanning the ops, scan the changes.