rust: Don't track encoding on ObjMeta
Context: the `automerge::types::ObjMeta` type tracks metadata about a
particular object. This type included an `encoding` field which tracks
how the OpTree should measure the width of operations in the object -
for lists this is always one, whereas for text objects this is the
number of utf-8 characters in the op.

Problem: The `ObjMeta::encoding` actually depends on the
`TextRepresentation` the caller supplies in various places (most
importantly the `PatchLog`), which makes it's presence in the `ObjMeta`
awkward because obtaining an `ObjMeta` now requires threading a
`TextRepresentation` through from the caller, and sometimes there is no
relevant `TextRepresentation` at all because we don't care about the
encoding.

Solution: remove the `ObjMeta::encoding` field and instead calculate the
encoding at the callsite using the callers `TextRepresentation`. In the
process we simplify some internal APIs as we can now remove
`Automerge::exid_to_just_obj` which was used when there was no supplied
`TextRepresentation` and we can refactor various parts of the codebase
which used a tuple of `(ObjId, ObjType)` to use the new `ObjMeta`.