rust: Add blocks related methods
This commit introduces the first components of rich text. To model rich
text in `ObjType::Text` objects we allow the insertion of
`ObjType::Map`s inline with the characters of the text. These maps
represent "block markers". This patch introduces the following new APIs
for working with block markers:

* Transactable::{split_block, join_block, update_block}
* ReadDoc::{spans, spans_at}

The first set of APIs offer ways to create, modify, and remove block
markers, whilst the second offer ways to view the state of the entire
text sequence.