Add Transactable::update_object
The introduction of blocks motivates an API I've been thinking of adding
for a while: the ability to update the state of some object by passing
in a new object which will be diffed with the current state. This
becomes more pressing with blocks because blocks are small maps stored
inline with the text and it's tedious to have to park object IDs
everywhere to manage these small maps; especially in Javascript where
there is no way of attaching metadata to a string.

`Transactable::update_block(object_id, value)` takes a
`automerge::hydrate::Value` which must be an object of the same type as
the type which `object_id` points at and figures out what set of changes
must be made to udpate the document to match the new value. This is
_not_ a very smart diff, it doesn't detect moves or modifications of
subobjects, but for small changes it should suffice.