Add Transactable::update_blocks
Problem: the block manipulation methods introduced in c17b6fb4d are
awkward to use with some text editors. The problem is that many text
editors expose a tree-structured data type, which is complicated to map
to the sequence of block markers and text which automerge users.

Solution: provide `Transactable::update_blocks`, which accepts a
sequence of `BlockOrText` objects and diffs these against the current
state of the text. This means that editors which provide drastically
different APIs need only provide a function which transforms the editor
state into the block list, rather than having to map the individual tree
mutations to block changes.