fix marks not appearing in some splice patches
Problem: inserting a character should emit a splice patch which includes
the marks active for the character which was just inserted, in some
situations this wasn't happening. The reason was that we have an
optimisation where each text object stores the position and op ID of the
last element that was inserted, this means that if you are inserting a
character immediately after one you just inserted (the common case when
typing) then we don't have to search the op tree for the insertion
point. We also use this cached information to generate the splice patch,
but that meant that we didn't know what marks to apply on the generated
patch.

Solution: also cache the marks active at the last seen character and add
them to the emitted splice patch.