automerge-wasm: add splitBlock,joinBlock,updateBlock,spans,updateSpans
This commit exposes the methods developed so far in `automerge` for
manipulating blocks. Most of the work is fairly mechanical but one
notable change is that in order to accept a sequence of POJOs as the
values for blocks it was necessary to extend the way "external types"
work. The problem is that we want users to be able to do things like
specify the value of a block attribute be a `RawString`, but
`automerge-wasm` has no understanding of the `RawString` class, instead
the `RawString` class is injected via the `registerDatatype` call, which
provides a function that should be called when _constructing_ an
instance of `RawString` (or `Counter`) to pass back to
`@automerge/automerge`. This means that `automerge-wasm` has no way of
recognising a `RawString` when it is passed in as part of a block.

For example, given this API call:

```
doc.updateSpans(<some object>, [
    {type: "block", value: {"key": new RawString("value")}}
])
```

We need to provide some way for `automerge-wasm` to see the `RawString`
type and convert it into a `ScalarValue::Str`. The solution I chose to
this was to add an argument to the `regsiterDatatype` call, which is
another function which "deconstructs" the data type, returning the base
value which should be inserted. So the `registerDatatype` call for
`RawString` looks like this:

```
registerDatatype(
    "str",
    (s) => new RawString(s),
    (s) => if (s instanceof RawString) s.val
)
```