js: update the materialized view even if it isn't patched
Context: the object which represents the document in the user facing API
is a POJO with some hidden attributes which we use to store various
metadata on - including a pointer to the opset. The functions which
operate on an automerge document lookup this metadata. Every time we
update the document (via e.g. Automerge.change) we apply any relevant
patches to the old object transforming it into a new object which we
return, then we mark the old object as stale.

Problem: when there were no patches to apply we were returning the old
object, but still marking it as stale. This was a little subtle because
there was actually a check for the "no patches" case, but not a check
for the "there were patches but none of them applied". This is a new
scenario introduced by blocks, because block patches don't have anything
to apply to in the materialized view (spans are only accessible via
`Automerge.spans`).

Solution: always update the object, even if no patches were applied.