fix: correct index calculations for large patch sets
Problem: when a change produced large numbers of patches (more than 100)
some patches could end up with incorrect indexes. The reason for this
was that the optimization which kicks in at > 100 patches which
precalculates the paths for every visible object in the document was
using the wrong encoding for calculating the indexes in lists.

Solution: use the correct encoding.