splice: fix no-expand marks at the end of text
Problem: in text which has a mark ending at the end of the text but
which was set to not expand on insertions at the end, insertions were
nevertheless still producting patches which contained the mark. The
reason was that the logic in the insert query for processing the
end-mark operations was only running when the next visible operation
_after_ the mark operations was processed.

Solution: run the mark processing logic when we encounter the end mark
operation.

While I was here I also added some exposition to the earlier test
exposing a problem with marks which cross optree boundaries and
generally cleaned that test up.