fix expand marks not appearing in remote patches
Problem: when splicing text at the end of a mark which is set to
ExpandMark::After the patches produced when merging the splice did not
contain the marks. The reason was that the `Opid` query unconditionally
runs each op it encounters through the marks state machine, but when it
is searching for an opid to insert after it ends up processing the
operation _following_ the opid it eventually returns, which meant the
marks state machine could process an end mark operation which is not
actually active at the insertion poitn.

Solution: only run an op through the marks state machine if it is not
greater than the op we are searching for.