Fix missing objects in reverse patches
Problem: when diffing "in reverse", i.e. where the start heads are
topologically after the end heads in the change graph, the patches
produced would not include objects which are "undeleted".

Solution: when traversing the OpSet if we encounter ops which make
create an object and which were visible before the heads we are starting
from, then mark the object as exposed so that the ExposeQueue will
render the object to the patch log later.