Fix corruption caused by missing actor indexes
Context: when we save a document we encode all the actor IDs in the
document at the start of the encoded document and then refer to those
actor IDs by their index into the encoded array. Similarly, in the
runtime representation of ops we don't directly store the actor ID but
instead refer to an index into a table of actor IDs held elsewhere. This
means that when we write the document to storage we need to translate
from the runtime indexes to the stored indexes.

Problem: in some cases we were creating actor indexes in the runtime
table of actors which we never wrote to the document. This meant that
the runtime actor indexes could get out of sync with the actors we write
to the document. This resulted in the actor indexes being wrong when
loading the document and consequently in "missing actor ID" errors.

Solution: when encoding actor IDs don't use the runtime table as a
translation table directly. Instead build a new table which contains
only the actor IDs we know we are writing.t