Reorganise WebAssembly packaging
Problem: there are a many JavaScript environments where it's impossible
to load WebAssembly modules as ES modules but where WebAssemly _is_
supported. This means Automerge requires a bunch of hacks to be usable
in these environments.

Solution: reorganise the WebAssembly packaging. Rather than the current
setup where we produce a low level `@automerge/automerge-wasm` package
from the `automerge-wasm` crate and depend on this package in
`@automerge/automerge`, we instead copy the WebAssembly output of
`wasm-bindgen` into the source of the `@automerge/automerge` package.
This gives us more control over how the WebAssembly is loaded and thus
allows us to expose an API for explicitly loading the WebAssembly. This
conditional loading is achieved via subpath exports and is explained in
depth in HACKING.md.

While we're here, we also reorganise the release process so that the NPM
package is published whenever a release is created rather than
automatically creating a release when the version number in package.json
changes.