rust: Add {Automerge, AutoCommit}::has_our_changes
Problem: it's often useful for the application to be able to tell
whether a remote peer has all the changes we have, but currently that
requires understanding the internal state of the automerge::sync::State
struct.

Solution: Add a method to Automerge and AutoCommit that returns true if
the replica represented by the sync state has all the changes we have.