js: Add topoHistoryTraversal
Problem: sometimes it's useful to be able to examine the history of the
document. At the moment the only way to do this is to use `getChanges`
and then `decodeChange`. This is cumbersome and slow.

Solution: in a first step towards more useful APIs introduce
`topoHistoryTraversal` which returns all of the change hashes in the
history in topological order. This is not a good API yet, so we put it
behind a @hidden annotation, but it's a start.