Add Automerge.stats
Problem: sometimes it is hard to know why performance is suffering for a
given document. Often, the issue is that the document history is large,
but it's not easy to examine this in a general manner.

Solution: Add a `Automerge.stats` function that returns a JSON object
indicating the number of changes and the number of ops in the document.