make getCursor work with Automerge.view
Problem: The `getCursor(doc, path, number)` - which returns a stable identifier
for a given index into a text object - always resolved the index with
respect to the latest heads of the document. In particular, if you
passed the result of `Automerge.view` as the first argument to
`getCursor` and then attempted to get a cursor for an index which had
since been deleted then a `RangeError` was thrown.

Solution: Respect the heads of the document passed to `getCursor` when
resolving the index.