rust: Pin to rustc 1.81
Problem: due to an intersection of changes in LLVM and wasm-bindgen
described here[1], compiling with rustc1 1.82.0 breaks the WebPack
webassembly integration. It looks like wasm-bindgen plan to push a fix
for this (despite it not really being their responsibility) but that's
not in yet.

Solution: pin the compiler to 1.81 until the wasm-bindgen fix is int.

[1]: https://github.com/rustwasm/wasm-bindgen/issues/4211