js: Make detecting Text and RawString more robust
Problem: when adding new values to a document in a `change` callback we
need to detect whether something is an instance of RawString or Text.
Currently we use an `instanceof` check. This can fail if the new value
was created in a different realm to the one where the check is being
executed. 

Solution: instead of using an instanceof check, use the presence of a
particular Symbol on the new values. The symbols are created using
Symbol.for() so they are shared across realms.