automerge-wasm: Pin version of wasm-bindgen
Problem: The wasm-bindgen dependency of automerge-wasm has to be exactly
the same as the version of the CLI used to build the JavaScript package.
The CI environment is pinned to a specific version in the Github actions
workflow but the version of wasm-bindgen in the Cargo.toml file is not. 
This means that when a new version of wasm-bindgen is released, the
build fails because cargo automatically pulls the latest version of
wasm-bindgen which is different to the version used in the CI
environment.

Solution: Pin the version of wasm-bindgen in the Cargo.toml to the same
version as in CI.