rust: allow specifying text encoding at document creation
Problem: when users perform operations like "insert at index <x> in text
<y>" Automerge has to transform the index into a position in the OpSet.
Different platforms prefer different encodings for text, so we need to
know what encoding to use to translate the index the user has given us
into a position in the OpSet. Currently this configuration is selected
by compile time feature flags, but in some cases users don't know what
encoding the platform they are targeting uses until runtime (this is the
case for Rust code which is compiled to WASM, but not always used in a
browser).

Solution: Make the text encoding a parameter at document creation.