new cursor API (#994)
The cursor API to date has worked like this: you call Automerge.getCursor(<doc>, <path to text>, <number>) to return a "cursor", which is a string representing the given position in such a way that you can later call Automerge.getCursorPosition(<cursor>) to lookup the position of the character you that was originally at index <number> after some changes. This has been useful but it has two shortcomings:

1. It is not possible to point at the beginning or the end of the text
2. If the character you were originally pointing at has been deleted it's not obvious what the result of getCursorPosition should be

This commit introduces two new things to the API to address this:

1. The ability to call Automerge.getCursor(<doc>, <path>, "start")  or Automerge.getCursor(<doc>, <path>, "end") to point at the beginning or end of the text
2. An optional additional argument to getCursor which determines how the cursor will be resolved when looking up it's position later. "before" indicates that the cursor should resolve to the first character preceding the target which was visible at the time the cursor was created and "after" resolves to the first character following the target.

There is one possible performance gotcha here. The implementation of the "before" resolution option scans through the whole text CRDT for each invisible character preceding the target until it finds a visible character. We don't actually know anyone who has a usecase for this behaviour though and are really including it for consistency. If it becomes a problem for someone we could introduce a new query type to seek backwards through the text, but this is a fair chunk of work so we defer until someone actually needs it.