Fix bug with `update_text` at old heads
`Transaction::update_text` implements a diff based method for modifying
text in an automerge document. The method accepts a new string and
performs a myers diff between the new string and the existing string in
order to determine the changes that need to be made to the text.

Calling `update_text` from within a transaction that is based on
non-current heads (such as within an `AutoCommit` on which you have
called `AutoCommit::isolate`) would erroneously calculate the diff
against the text from the latest heads, rather than the heads the
transaction was based on. This commit fixes the issue by using the
transaction heads to materialize the text to diff against.