Don't show removed marks in AutoCommit::get_marks
Problem: The `AutoCommit::get_marks` call which returns the marks active
for a particular index in the text returns a map from mark name to
value. This map still contains marks that have been removed, just with
their value set to `ScalarValue::Null`. This is inconsistent with the
`marks` API and `spans` APIs which return the marks active for the whole
sequence of text and do not include removed marks at all.

Solution: Update the `AutoCommit::get_marks` function to only return
marks that are active for the given index.