Fix marks in spans not respecting heads
Problem: the spans returned by `Automerge::spans` always rendered the
"latest" marks in the document - i.e. the marks corresponding to the
tips of the commit graph. This means that if you call
`Automerge::spans_at(<some heads>)` (or, in JS,
`Automerge.spans(Automerge.view(doc, heads), <path>))` you get the
correct block markers (the block markers as at the given heads), but
incorrect marks (the marks as at the latest heads in the document). The
reason this was happening is that in the spans iterator, which iterates
over every op in the document, we correctly skip invisible ops when
processing blocks, but not when processing marks.

Solution: correctly skip invisible ops when processing marks.