js: Update webpack packaging tests to webpack 5.99.0
Problem: when using rustc > 1.82 and webpack < 5.97.0 the wasm file
generated by wasm-bindgen can't be loaded by webpack. This is tracked by
an issue [1] in the wasm-bindgen repo and an issue [2] in the webpack
repo. This meant that the packaging tests were failing.

Solution: upgrade the packaging tests to use webpack 5.99.0. This does
mean that people using older versions of webpack will need to upgrade,
but it's a patch version upgrade so shouldn't be too disruptive.

[1]: https://github.com/rustwasm/wasm-bindgen/issues/4211
[2]: https://github.com/webpack/webpack/issues/15566