js: denextify
For a long time we have had two APIs for the JavaScript
@automerge/automerge package. The "stable" API and the "next" API. The
"next" API had some significant differences from the stable API but we
didn't want to make a breaking change in the API until we were confident
of these differences. The most significant difference being the use of
`string` to represent text rather than a the `Text` class in the
"stable" API. We're now very sure of the "next" API and the complexity
of maintaining two APIs is no longer worth it, we're going to do a major
version bump to release the new compressed memory OpSet, so removing the
old "stable" API seems like a good thing to do at the same time.