js: Rename RawString to ImmutableString
Problem: RawString is an unhelpful name. A RawString is a string which
is not collaborative, we don't create an op for each character in it but
instead have a single op which sets the value of the whole string. It's
useful when you expect to have a large string which will never be edited
by a user (such as the output of a compilation step).

Solution: Rename RawString to ImmutableString, which better reflects
why you might use it.