js: Reintroduce the `next` export
In 582dea16 we removed the `next` export and subpath export. This was a
little premature as we want automerge-repo to be able to support
applications using both Automerge 2.0 and the upcoming Automerge 3.0
beta until we have a stable release of 3.0. To accomodate this we
re-introduce the `next` export, but not the subpath export. This
required moving the implementation out of `index.ts` and into
`implementation.ts` which can then be rexported twice from `index.ts`.