CI: hardcode rust toolchain
Problem: The rust toolchain in use in CI is set to "stable". This can
cause issues if the current stable toolchain is different to the version
in rust/rust-toolchain.toml. For example, `cargo fmt` will sometimes
complain that the `rustfmt` component is not installed.

Solution: set the toolchain in `.github/workflows/ci.yaml` to the same
as the version in `rust/rust-toolchain.toml`. We'll need to keep them in
sync but at least it's predictable.