js: Fix indexOf not finding string elements
Problem: the indexOf method which is exposed on lists inside of a
`change` callback does not handle strings correctly. For example, this
test

```
let doc = Automerge.from({list: ["a", "b", "c"]})
doc = Automerge.change(doc, d => {
    assert.equal(d.list.indexOf("b"), 1)
})
```

Will fail. The `indexOf` will return `-1`. The reason is that the
elements of the list are represented as text objects but we compare them
with native javascript strings inside `indexof`.

Solution: if the search element argument to `indexOf` is a string, check
whether the candidate element is a text object and compare the search
element to the reified text of the candidate element. I also commented
the implementation as it was a little terse and hard to follow.