automerge-wasm: use autogenerated typescript bindings
To date we have been manually writing typescript bindings for
the wasm-bindgen generated code. This was because some parts of the
interface we provide can't be typed using the wasm_bindgen macro. This
is not ideal though because it becomes much easier for the typescript
types to diverge from the implementation. Recent additions to
wasm_bindgen[1] allow us to generate more typescript bindings
automatically.

There are still some functions which we can't generate using this new
functionality, but we can get around that by using typescript
declaration merging. We declare an interface in the typescript custom
section[2] which merges with the class declaration generated by
wasm_bindgen.

This commit implements the exact same interface as the manually written
one but with much of the types autogenerated from the rust types. This
should make it easier to maintain going forwards.

[1]: https://github.com/rustwasm/wasm-bindgen/pull/4394
[2]: https://rustwasm.github.io/wasm-bindgen/reference/attributes/on-rust-exports/typescript_custom_section.html