js: freeze getConflicts return value outside of change callbacks
Problem: the getConflicts call returns proxies which allow users to
modify the values inside conflicted values. This is great inside a
change callback but outside of a change callback it allows modifications
of a document by modifying the conflicted values. This is extremely
confusing as it's entirely opposite to how the rest of the API works.

Solution: detect if we are inside a change callback in `getConflicts`
and if we are not, return a frozen POJO object instead of a proxy. This
is much more predictable. 

As a side benefit, this fixes some tests which recently started failing
due to changes in the way `assert.deepStrictEqual` treats proxies in
recent node versions.