rust: fix noop handling
Problem: when creating a new operation we attempt to check whether the
new operation is a "noop" and if it is, we don't do anything. One place
where this is a little subtle is that if the new operation is being
created only to resolve a conflict. In this case the value is the same
as the existing value in the document. We correctly identify this case
and emit an operation, however, the operation we emit overrides all the
previous values. This can lead to a scenario where concurrent nodes get
into a loop attempting to resolve the conflict. For example, imagine
node A and B are collaborating on a document.

* Node A sets key "foo" to "bar"
* Node B sets key "foo" to "baz"
* The two nodes send changes to each other
* Node A sees that key "foo" is conflicted and sets it to "qux"
* Node B also sees that key "foo" is conflicted and also sets it to
  "qux"
The two nodes send changes to each other
* Node A again sees that the key "foo" is conflicted, and so again sets
  it to "qux"
* Node B also sees that the key "foo" is conflicted, and so again sets
  it to "qux"
The above loops indefinitely

The reason for the loop is the fact that both A and B see the "foo" bar
as conflicted because the op they create to resolve the conflict always
overrides all the previous conflicted values.

Solution: instead of overriding all previous values when resolving a
conflict, we delete the conflicting values. The winning value is chosen
as per the normal automerge rules of the op with the lowest lamport
timestamp. All nodes will eventually agree on which timestamp this is
and thus the loop will not occur.