rust: getrandom v3 and rand v0.9
Most of this commit is a mechanical set of changes to upgrade rand and
getrandom to their respective latest versions. There is one irritating
wrinkle. getrandom v0.3 has removed a polyfill for the webcrypto API in
Node 18 CommonJS environments. To avoid breaking anyone still using
automerge in such an environment I've had to add this polyfill to our
own codebase. The excruciating details are in
`javascript/HACKING.md#getrandom-support`.