rust: fix a bug where marks didn't cross block markers in spans
Problem: the `Automerge::spans` function didn't correctly report spans
with marks which crossed block markers. This was quite fiddly to debug
because the logic for emitting spans is a bit complex.

Solution: clean up the spans iterator a bit so it's clearer when we're
emitting spans, then always keep track of the active marks so that when
we emit a block marker we don't lose the state of marks which cross the
block marker.