rust: update marks in update_spans
The `Transactable::update_spans` method performs a myers diff over the
text in a document and tries to performa a minimal set of updates to the
document in order to make the document content the same as the spans
passed in. This function didn't update the marks, this was extremely
surprising to most people.

The main reason marks were not updated is that I wasn't sure how to
handle the creation of the `ExpandMark` flags on marks which are created
as part of the diff reconciliation. The difficulty is that the expand
flag is not part of the span state, for example, if there are multiple
overlapping "bold" marks in a document we only report a single bold
span. This means we don't have an obvious place to put the "expand" flag
in the input to `update_spans`. After some thought I have decided that
the best way to handle this is to add a `config` parameter to
`update_spans` which controls how the expand flag for new marks is
created. This is not as dynamic as allowing per-span expand flags, but
it is a lot simpler and reflects the fact that most applications set the
expand behavior for all marks of a given type in the same way.

With this change we are able to remove the `BlockOrText` type, which was
only different to the `Span` type in that it did not have marks on it.
Now that we accept marks in `update_spans` we can use the `Span` type as
the output of `ReadDoc::spans` and as the input to
`Transactable::update_spans`.